Add .gitignore, create index.php with book player content
Created .gitignore file to exclude .idea directory, metadata and .mp3 files. Simultaneously, an index.php file was added with complete layout and functionality for a book player. This includes php logic for handling JSON metadata and dynamic generation of mp3 elements as well as extensive styling and scripting for the interactive features.